refactor(suite-native): backup recap components made reusable - these components will be reused in wallet recovery recap
