+++ suite-native/module-device-onboarding/src/components/WalletBackupRecap/create/WalletBackupRecapStep1.tsx
@@ -6,7 +6,7 @@ import { prepareNativeStyle, useNativeStyles } from '@trezor/styles';
 
 import { WalletBackupRecapListItem } from './WalletBackupRecapListItem';
 import { WALLET_BACKUP_RECAP_STEPS, walletBackupSecuritySteps } from './presets';
-import { SwipeableWalkthroughStep } from '../SwipeableWalkthrough/SwipeableWalkthroughStep';
+import { SwipeableWalkthroughStep } from '../../SwipeableWalkthrough/SwipeableWalkthroughStep';
 
 export type WalletBackupTutorialNumberedStepProps = {
     currentStepIndex: SharedValue<number>;
