+++ suite-native/module-device-onboarding/src/components/WalletBackupRecap/create/presets.ts
