+++ suite-native/module-device-onboarding/src/screens/WalletBackupRecapScreen.tsx
@@ -1,4 +1,3 @@
-import { useCallback } from 'react';
 import { useSharedValue } from 'react-native-reanimated';
 
 import { Screen } from '@suite-native/navigation';
@@ -6,32 +5,24 @@ import { Screen } from '@suite-native/navigation';
 import { SwipeableWalkthrough } from '../components/SwipeableWalkthrough/SwipeableWalkthrough';
 import { SwipeableWalkthroughScreenHeader } from '../components/SwipeableWalkthrough/SwipeableWalkthroughScreenHeader';
 import { WalletBackupRecapBackButton } from '../components/WalletBackupRecap/WalletBackupRecapBackButton';
-import { WalletBackupRecapStep1 } from '../components/WalletBackupRecap/WalletBackupRecapStep1';
-import { WalletBackupRecapStep2 } from '../components/WalletBackupRecap/WalletBackupRecapStep2';
-import { WalletBackupRecapStep3 } from '../components/WalletBackupRecap/WalletBackupRecapStep3';
-import { WalletBackupRecapStep4 } from '../components/WalletBackupRecap/WalletBackupRecapStep4';
-import { WALLET_BACKUP_RECAP_STEPS } from '../components/WalletBackupRecap/presets';
+import { WalletBackupRecapStep1 } from '../components/WalletBackupRecap/create/WalletBackupRecapStep1';
+import { WalletBackupRecapStep2 } from '../components/WalletBackupRecap/create/WalletBackupRecapStep2';
+import { WalletBackupRecapStep3 } from '../components/WalletBackupRecap/create/WalletBackupRecapStep3';
+import { WalletBackupRecapStep4 } from '../components/WalletBackupRecap/create/WalletBackupRecapStep4';
+import { WALLET_BACKUP_RECAP_STEPS } from '../components/WalletBackupRecap/create/presets';
 import { useExitAlert } from '../hooks/useExitAlert';
 
 export const WalletBackupRecapScreen = () => {
     const currentStepIndex = useSharedValue(0);
     const { handleExitButtonPress } = useExitAlert();
 
-    const handlePressBackButton = useCallback(() => {
-        if (currentStepIndex.value === 0) {
-            handleExitButtonPress();
-        } else {
-            currentStepIndex.value -= 1;
-        }
-    }, [currentStepIndex, handleExitButtonPress]);
-
     return (
         <Screen
             header={
                 <SwipeableWalkthroughScreenHeader
-                    onPressBack={handlePressBackButton}
                     currentStepIndex={currentStepIndex}
                     CustomBackButton={WalletBackupRecapBackButton}
+                    onPressBack={handleExitButtonPress}
                 />
             }
             isScrollable={false}
